Refactor linux timing into separate file
