+++ time-driver.h
@@ -1,0 +1,5 @@
+#include "ssm-queue.h"
+
+ssm_time_t timestep();
+
+void initialize_time_driver();
